Charger Research (11/20/2023)
Includes background research on EVSEs, regulations affecting the charger itself, and research on the Elcon HK-J 6.6kW charger (most common charger). Further research is on pause until confirmation of battery pack/accumulator used, and other information to determine if Elcon products fit the needs, or if another product is required (or if the charger itself is also to be built from scratch by the team).